backend/requirements.txt
fastapi
uvicorn
msal
requests
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
openai
httpx
python-multipart